requirements.txt
openai
langchain
langchain-openai
langchain-community
faiss-cpu
pandas
numpy
matplotlib
seaborn
plotly
streamlit
python-dotenv
tiktoken
